requirements.txt
anthropic
python-docx
pyodbc
python-dotenv
